move common form classes to common/forms.py
