+++ mep/common/forms.py
@@ -4,6 +4,46 @@ from django.core.validators import RegexValidator
 from django.utils.text import mark_safe
 
 
+class SelectDisabledMixin():
+    '''
+    Mixin for :class:`django.forms.RadioSelect` or :class:`django.forms.CheckboxSelect`
+    classes to set an option as disabled. To disable, the widget's choice
+    label option should be passed in as a dictionary with `disabled` set
+    to True::
+
+        {'label': 'option', 'disabled': True}.
+    '''
+
+    # Using a solution at https://djangosnippets.org/snippets/2453/
+    def create_option(self, name, value, label, selected, index, subindex=None,
+                      attrs=None):
+        disabled = None
+
+        if isinstance(label, dict):
+            label, disabled = label['label'], label.get('disabled', False)
+        option_dict = super().create_option(
+            name, value, label, selected, index,
+            subindex=subindex, attrs=attrs
+        )
+        if disabled:
+            option_dict['attrs'].update({'disabled': 'disabled'})
+        return option_dict
+
+
+class RadioSelectWithDisabled(SelectDisabledMixin, forms.RadioSelect):
+    '''
+    Subclass of :class:`django.forms.RadioSelect` with option to mark
+    a choice as disabled.
+    '''
+
+
+class SelectWithDisabled(SelectDisabledMixin, forms.Select):
+    '''
+    Subclass of :class:`django.forms.Select` with option to mark
+    a choice as disabled.
+    '''
+
+
 class CheckboxFieldset(forms.CheckboxSelectMultiple):
     '''Override of :class:`~django.forms.CheckboxSelectMultiple`
     that renders as a fieldset with checkbox inputs.'''
